# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Ride-With-Me",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/notifications/get/{user}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get user notifications",
                "description": "Retrieve all notifications for a specific user.",
                "operationId": "getUserNotifications",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Notification"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/report/reasons/get": {
            "get": {
                "tags": [
                    "ReportReasons"
                ],
                "summary": "Get all report reasons",
                "description": "Retrieve a list of all report reasons.",
                "operationId": "getReportReasons",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReportReason"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Notification": {
                "title": "Notification",
                "description": "Notification model",
                "required": [
                    "id",
                    "user_id",
                    "sender_id",
                    "type",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "description": "Primary key of the notification",
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "description": "ID of the user who receives the notification",
                        "type": "string",
                        "format": "uuid"
                    },
                    "sender_id": {
                        "description": "ID of the user who sent the notification",
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "description": "Type of the notification",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp when the notification was created",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ReportReason": {
                "title": "ReportReason",
                "description": "ReportReason model",
                "required": [
                    "id",
                    "reason"
                ],
                "properties": {
                    "id": {
                        "description": "Primary key of the report reason",
                        "type": "string",
                        "format": "uuid"
                    },
                    "reason": {
                        "description": "Description of the report reason",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}